requirements.txt
dlib==19.17.99
entrypoints==0.3
face-recognition==1.2.3
face-recognition-models==0.3.0
fbs==0.8.3
imutil==0.2.5
imutils==0.5.2
matplotlib==3.1.1
mysql-connector-python==8.0.16
numpy==1.16.4
opencv-contrib-python==4.1.0.25
Pillow==6.1.0
PyPDF2==1.26.0
PyQt5==5.13.0
PyQt5-sip==4.19.18
reportlab==3.5.23
scikit-image==0.15.0
scipy==1.3.0
